fix: resolve cognitive tunneling and align studio foundations
This commit addresses the "circular reasoning" and "cognitive tunneling" failures identified in SESSION-01 by implementing a fundamentally different approach focused on foundational alignment:

1. **Guardrail Calibration**: Lowered the Semantic Entropy (SE) threshold from 7.0 to 2.0 system-wide (memory, code, and docs) to match the mathematical limits of Shannon Entropy for N=5 samples, making the circuit breaker significantly more sensitive.
2. **Structural Alignment**: Moved `OptimizerAgent` to `studio/agents/optimizer.py` and implemented the `apply_optimizations` method required by the Orchestrator for OPRO (Optimization by PROmpting).
3. **Environment Repair**: Updated `requirements.txt` with missing core dependencies (`langgraph`, `langchain-core`, etc.) and resolved 28 test collection errors.
4. **Model Stratification**: Updated `ProductOwnerAgent` to use central settings for model selection, eliminating hardcoded invalid model strings.
5. **Documentation Correction**: Refactored `README.md` to correctly describe agent roles (Engineer as Proxy vs QA Verifier in Sandbox) and synchronized versioning/threshold information.

Verified by running the full 108-test suite (106 passed, 2 skipped).
