+++ requirements.txt
@@ -1,10 +1,16 @@
 langchain
+langchain-community
+langchain-core
+langgraph
 langchain-google-vertexai
+langchain-chroma
+google-cloud-aiplatform
+pypdf
 pydantic
 pytest
+pytest-asyncio
 PyGithub
 python-dotenv
-langchain-google-vertexai
 docker
 networkx
 pydantic-settings
