+++ tests/test_context_slicing.py
@@ -50,7 +50,7 @@ def test_context_slicing_large_log(mock_run_retrospective, mock_run_po, mock_gen
     # Mock calculator
     mock_metric = SemanticHealthMetric(
         entropy_score=0.5,
-        threshold=7.0,
+        threshold=2.0,
         sample_size=5,
         is_tunneling=False,
         cluster_distribution={}
