+++ tests/test_escalation.py
@@ -73,7 +73,7 @@ async def test_escalation_logic():
 
         low_entropy_metric = SemanticHealthMetric(
             entropy_score=0.5,
-            threshold=7.0,
+            threshold=2.0,
             sample_size=5,
             is_tunneling=False,
             cluster_distribution={"Cluster_0": 1.0}
