+++ tests/test_lifecycle_manager.py
@@ -47,7 +47,7 @@ def test_orchestrator_lifecycle_manager(mock_run_retrospective, mock_run_po, moc
     orchestrator = Orchestrator(engineer_app=mock_engineer_app)
     mock_metric = SemanticHealthMetric(
         entropy_score=0.5,
-        threshold=7.0,
+        threshold=2.0,
         sample_size=5,
         is_tunneling=False,
         cluster_distribution={}
