+++ tests/test_max_retry.py
@@ -63,7 +63,7 @@ async def test_max_retry_containment():
 
         low_entropy_metric = SemanticHealthMetric(
             entropy_score=0.5,
-            threshold=7.0,
+            threshold=2.0,
             sample_size=5,
             is_tunneling=False,
             cluster_distribution={"Cluster_0": 1.0}
