+++ tests/test_refactor_loop.py
@@ -79,7 +79,7 @@ async def test_refactor_loop():
     # Mock Entropy Calculator to return LOW ENTROPY
     low_entropy_metric = SemanticHealthMetric(
         entropy_score=0.5,
-        threshold=7.0,
+        threshold=2.0,
         sample_size=5,
         is_tunneling=False,
         cluster_distribution={"Cluster_0": 1.0}
