+++ tests/test_router_diversion.py
@@ -44,7 +44,7 @@ async def test_router_case_a_coding_with_log(mock_gen_model, mock_vertex_judge):
     # Mock the calculator to avoid network calls
     mock_metric = SemanticHealthMetric(
         entropy_score=0.5,
-        threshold=7.0,
+        threshold=2.0,
         sample_size=5,
         is_tunneling=False,
         cluster_distribution={}
